# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个使用 Swagger 进行接口文档管理的 Go API 服务器。",
        "title": "Gotribe Admin API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/api",
    "paths": {
...
        "/post": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有内容的列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容管理"
                ],
                "summary": "获取内容列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "postID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "projectID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容管理"
                ],
                "summary": "创建内容",
                "parameters": [
                    {
                        "description": "创建内容请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.CreatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据内容ID列表批量删除内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容管理"
                ],
                "summary": "批量删除内容",
                "parameters": [
                    {
                        "description": "删除内容请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.DeletePostsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/post/{postID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据内容ID获取内容详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容管理"
                ],
                "summary": "获取内容信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "内容ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据内容ID发布内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容管理"
                ],
                "summary": "发布内容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "内容ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据内容ID更新内容信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容管理"
                ],
                "summary": "更新内容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "内容ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.UpdatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "description": "统一API响应格式",
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "操作成功"
                }
            }
        },
...
        "vo.CreatePostRequest": {
            "type": "object",
            "required": [
                "author",
                "categoryID",
                "content",
                "description",
                "htmlContent",
                "projectID",
                "title",
                "type",
                "userID"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "categoryID": {
                    "type": "string"
                },
                "columnID": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 300,
                    "minLength": 2
                },
                "ext": {
                    "type": "string"
                },
                "htmlContent": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isPasswd": {
                    "type": "integer"
                },
                "isTop": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "people": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "type": {
                    "type": "integer"
                },
                "unitPrice": {
                    "type": "number"
                },
                "userID": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
        "vo.DeletePostsRequest": {
            "type": "object",
            "properties": {
                "postIds": {
                    "type": "string"
                }
            }
        },
...
        "vo.UpdatePostRequest": {
            "type": "object",
            "required": [
                "author",
                "categoryID",
                "content",
                "description",
                "htmlContent",
                "projectID",
                "title",
                "type",
                "userID"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "categoryID": {
                    "type": "string"
                },
                "columnID": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 300,
                    "minLength": 2
                },
                "ext": {
                    "type": "string"
                },
                "htmlContent": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isPasswd": {
                    "type": "integer"
                },
                "isTop": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "people": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "type": {
                    "type": "integer"
                },
                "unitPrice": {
                    "type": "number"
                },
                "userID": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请输入 Bearer {token}，注意 Bearer 和 token 之间有一个空格",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}